Consolodate Kibana and Elasticsearch launch logic in tests (#22328)
