+++ src/test_utils/kbn_server.js
@@ -20,8 +20,9 @@
 import { resolve } from 'path';
 import { defaultsDeep, set } from 'lodash';
 import { header as basicAuthHeader } from './base_auth';
-import { esTestConfig, kibanaTestUser, kibanaServerTestUser } from '@kbn/test';
+import { createEsTestCluster, esTestConfig, kibanaTestUser, kibanaServerTestUser } from '@kbn/test';
 import KbnServer from '../../src/server/kbn_server';
+import { ToolingLog } from '@kbn/dev-utils';
 
 const DEFAULTS_SETTINGS = {
   server: {
@@ -67,10 +68,58 @@ export function createServer(settings = {}) {
 }
 
 /**
+ * Creates an instance of KbnServer, including all of the core plugins,
+ * with default configuration tailored for unit tests, and starts es.
+ *
+ * @param  {Object} options
+ * @prop {Object} settings Any config overrides for this instance
+ * @prop {function} adjustTimeout A function(t) => this.timeout(t) that adjust the timeout of a test,
+ *    ensuring the test properly waits for the server to boot without timing out.
+ * @return {KbnServer}
+ */
+export async function startTestServers({ adjustTimeout, settings = {} }) {
+  if (!adjustTimeout) {
+    throw new Error('adjustTimeout is required in order to avoid flaky tests');
+  }
+
+  const log = new ToolingLog({
+    level: 'debug',
+    writeTo: process.stdout
+  });
+
+  log.indent(6);
+  log.info('starting elasticsearch');
+  log.indent(4);
+
+  const es = createEsTestCluster({ log });
+
+  log.indent(-4);
+
+  adjustTimeout(es.getStartTimeout());
+
+  await es.start();
+
+  const kbnServer = createServerWithCorePlugins(settings);
+
+  await kbnServer.ready();
+  await kbnServer.server.plugins.elasticsearch.waitUntilReady();
+
+  return {
+    kbnServer,
+    es,
+
+    async stop() {
+      await this.kbnServer.close();
+      await es.cleanup();
+    },
+  };
+}
+
+/**
  *  Creates an instance of KbnServer, including all of the core plugins,
  *  with default configuration tailored for unit tests
  *
- *  @param  {Object} [settings={}]
+ *  @param  {Object} [settings={}] Any config overrides for this instance
  *  @return {KbnServer}
  */
 export function createServerWithCorePlugins(settings = {}) {
