+++ src/ui/ui_settings/routes/__tests__/lib/servers.js
@@ -17,39 +17,24 @@
  * under the License.
  */
 
-import { createEsTestCluster } from '@kbn/test';
-import { ToolingLog } from '@kbn/dev-utils';
-import * as kbnTestServer from '../../../../../test_utils/kbn_server';
+import { startTestServers } from '../../../../../test_utils/kbn_server';
 
 let kbnServer;
 let services;
-let es;
+let servers;
 
 export async function startServers() {
-  const log = new ToolingLog({
-    level: 'debug',
-    writeTo: process.stdout
-  });
-  log.indent(6);
-
-  log.info('starting elasticsearch');
-  log.indent(4);
-
-  es = createEsTestCluster({ log });
-  this.timeout(es.getStartTimeout());
-
-  log.indent(-4);
-  await es.start();
-
-  kbnServer = kbnTestServer.createServerWithCorePlugins({
-    uiSettings: {
-      overrides: {
-        foo: 'bar',
-      }
+  servers = await startTestServers({
+    adjustTimeout: (t) => this.timeout(t),
+    settings: {
+      uiSettings: {
+        overrides: {
+          foo: 'bar',
+        }
+      },
     }
   });
-  await kbnServer.ready();
-  await kbnServer.server.plugins.elasticsearch.waitUntilReady();
+  kbnServer = servers.kbnServer;
 }
 
 export function getServices() {
@@ -57,7 +42,7 @@ export function getServices() {
     return services;
   }
 
-  const callCluster = es.getCallCluster();
+  const callCluster = servers.es.getCallCluster();
 
   const savedObjects = kbnServer.server.savedObjects;
   const savedObjectsClient = savedObjects.getScopedSavedObjectsClient();
@@ -78,14 +63,8 @@ export function getServices() {
 
 export async function stopServers() {
   services = null;
-
-  if (kbnServer) {
-    await kbnServer.close();
-    kbnServer = null;
-  }
-
-  if (es) {
-    await es.cleanup();
-    es = null;
+  kbnServer = null;
+  if (servers) {
+    await servers.stop();
   }
 }
